+++ lib/docx/formatting/paragraph_formatting.rb
@@ -5,6 +5,22 @@ module Docx
     include Formatting
 
     def apply_formatting(formatting)
+      formatting ||= {}
+
+      # 设置段落属性 -- 缩进
+      if (formatting[:ind])
+        ind_node = add_property('ind')
+
+        # ind_node["w:left"] = "0"
+        # ind_node["w:leftChars"] = "0"
+        # ind_node["w:firstLine"] = "0"
+        # ind_node["w:firstLineChars"] = "0"
+        # <w:ind w:left="420" w:leftChars="0" w:firstLine="411" w:firstLineChars="196" />
+        formatting[:ind].each do |key, value|
+          ind_node["w:#{key}"] = value
+        end
+      end
+
       if (formatting[:alignment])
         alignment_node = add_property('jc')
         alignment_node['w:val'] = formatting[:alignment]